requirement.txt
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.5
python-dotenv==1.0.1
gradio==5.13.0
pypdf==5.1.0
docx2txt==0.8
opencv-python==4.11.0.86
pytesseract==0.3.13

